Create SharingViewControllerViewsBuilder
